Refactor of the import
