+++ Platform/Windows/callbacks.go
@@ -3,7 +3,7 @@ package Windows
 import (
 	common "github.com/Erickype/GoGameEngine/Common"
 	"github.com/Erickype/GoGameEngine/Events"
-	"github.com/go-gl/glfw/v3.3/glfw"
+	"github.com/go-gl/glfw/v3.2/glfw"
 )
 
 func declareCallbacks(w *Window) {
